debug ver.
