+++ .fvmrc
@@ -1,4 +1,4 @@
 {
-  "flutter": "3.7.0",
+  "flutter": "3.24.3",
   "flavors": {}
 }
