+++ foodb/lib/src/key_value/key_value_get.dart
@@ -24,7 +24,7 @@ mixin _KeyValueGet on _AbstractKeyValue {
     }
     var entry = await keyValueDb.get(baseKey);
     if (entry == null) {
-      throw AdapterException(error: "not_found", reason: "missing");
+      throw AdapterException(error: "not_found", reason: ("missing key" + baseKey.toString()));
     }
     var result = DocHistory.fromJson(entry.value);
     Doc<T>? doc;
